add dialog loader
